requirements.txt
lido==1.1.0
click8==8.0.1
colorama==0.4.4